requirements.txt
pytest
request
pymysql
pyyaml